refactor: page size 상수 선언
