+++ src/main/java/gift/util/Utils.java
@@ -1,9 +1,9 @@
-package gift.product.util;
+package gift.util;
 
 public final class Utils {
     public static final String NAME_PATTERN = "^[ㄱ-ㅎ가-힣a-zA-Z0-9 ()\\[\\]\\+\\-&/_]*$";
 
-    public static final int PAGE_SIZE = 10;
+    public static final int DEFAULT_PAGE_SIZE = 10;
 
     private Utils() {
     }
